backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
email-validator==2.1.0

# Testing dependencies
pytest==8.0.0
pytest-flask==1.3.0
pytest-cov==4.0.0